requirements.txt
mlx
transformers
numpy
huggingface-hub
llama-cpp-python